it "works"
This is definitely technically working, but not the way i ultimately want this to stay. For now i just want to be able to see something, and everything else can follow.

+++ CombatlogParser/src/Data/DamageSummary.cs
@@ -9,7 +9,7 @@ namespace CombatlogParser.Data
     public class DamageSummary
     {
         public string SourceName { get; set; }
-        public uint TotalDamage { get; set; }
+        public long TotalDamage { get; set; }
         public float DPS { get; set; }
     }
 }
